More work to migrate to Bytecode DSL. Commits have been squashed to simplify rebase efforts. The original commit messages are:     Get the DSL to generate code without errors     Get the generated code to compile without errors     re-add code to enable Operations interpreter     Fix rebase issues + start factoring common validation out to AbstractCompiler.java     Fix kwarg validation     Fix implementation of Try     Add named exception unbinding     Fix implementation of With     Minor fixes to f-strings and for-else blocks     Refactor to support nested root node compilation without FunctionFinder     Begin refactoring generators     Partly fix generator implementation     Support plain one-way generators     Fix bug with empty generator expressions     Support generator send (non-exceptional cases)     Fully support enhanced generators     Remove intermediate root node in generator implementation     Add branch profile to generator resume code     Fix prepareCall to handle generator calling convention     merge generator logic back into a single PGenerator class     Fix argument validation     Fix unbound local checking and qualified names for functions     Encapsulate compiler state in an OperationCompilerContext     Support async def, async with, and await     Get more JUnit tests running with the Operation interpreter     Minor improvements to ResumeGenerator code     Get JUnit tests using only the Operation interpreter     Set catching frame reference for try-finally and with statements     Fix stack walking to support continuation frames     Minor location + integer constant fixes     Convert FrameInfo to an interface and provide separate impls for bytecode and operation interpreters     Make CodeUnit an abstract class with Bytecode and Operation implementations     Allocate locals contiguously from slot 0 & avoid redundant allocations     Misc. bug fixes (incl. name mangling and unbound local checking)     Fix class cell reads and docstrings     Fix bugs with keyword arguments + augmented assignments     Remove AbstractCompiler abstraction, add more syntax checks     Support type annotations     Implement co_consts and other code fields     Respect 'optimize' flag for compilations, constant fold unary negation     Fix constants implementation + parser bugs     Set exception __context__ when it is raised     Properly handle __future__ imports     Convert host StackOverflowError to RecursionError     Use findBci to compute frame.f_lineno     Use interceptTruffleException to set catchBci for exceptions     fix Operation DSL API usage     Move exception chaining to intercept hook     Make future annotations inherited by exec     Support interop method calls     Fix finallyTry, local accesses, and continuation API usages after DSL updates     Save + restore exception state on generator exit + entry     Use getLocals/copyLocals APIs instead of hard-coded local offset     Store bci into frame when potentially escaping root node     use Proxyable for OperationProxy nodes     Fix test setup after rebase     Update with operation->bytecode package rename     Rename POperationXYZ -> PBytecodeDSLXYZ     Refactor compiler to return a CodeUnit and use a unique Builder for each root node     Rework Marshal to use DataOutput and DataInput APIs     Basic serialization implementation     Avoid serialize+deserialize step for AST parsing     Only build root node once for each MakeFunction/Generator/Coroutine operation     Enable frozen modules, don't actually serialize Sources     Cleanup style errors.     Fix source creation, unintentional root node adoption, and parsing bugs     Fix bciToLine     Always use character-based sources     Add separate DSL config for tests     bump CI overlay     bump overlay, dsl -> bytecode-dsl in gate tags     peg imports to bytecode dsl branch     Fix some compilation issues and undo temporary test hacks     Break constant collection creation into separate operations     Re-enable pattern matching tests and start a DSL implementation     Add support for sequence pattern matching     Move EnableBytecodeDSLInterpreter from an option to a system property; fix NI compilation     bump overlay, add SVM DSL unittests config     (wip) attempt to create separate native standalone for the DSL interpreter     Pass bytecode DSL flag to builder when building standalone     disable fail fast for DSL tests     Fix inplace shifts; interactive string expressions     Fix bug with empty module __init__ files     Fix class decorators, disable pdb test on DSL interpreter, re-enable traceback test     Fix super() implementation & fix bug with shadowed cell/free variables     Update short-circuit ops, fix super() generator frame lookup     Correctly mangle kwargs; implement co_lines; fix imod implementation     Parse __future__ annotations before scope analysis to ensure bad annotations rejected     Skip boolean coercion for boolean constants     Use getLocal API for readClassCell/readSelf     Clean up marshal code     Use a PBytecodeDSLGeneratorFunctionRootNode instead of creating a bytecode node to instantiate generators     Remove extra import     Fix NI compilation by putting SourceSection method behind boundary; compute internal the same way as PBytecodeRootNode     Fix rebase issue + manual bc frozen module issue     bump ci overlay     Fix another rebase issue with mx configuration     bump overlay     update to use new Bytecode DSL APIs (wip)     remove bciToLine(int), set bytecode node when intercepting Truffle exceptions     Replace instanceof checks with PythonOptions.ENABLE_BYTECODE_DSL_INTERPRETER checks     getSourceLocation -> findSourceLocation     Undo PythonTests.newBuilder now that DSL is enabled by a system property     remove unreachable PKeyword marshalling logic

+++ graalpython/com.oracle.graal.python.frozen/freeze_modules.py
@@ -156,7 +156,7 @@ else:
 #######################################
 # specs
 
-def parse_frozen_specs():
+def parse_frozen_specs(suffix):
     seen = {}
     for section, specs in FROZEN:
         parsed = _parse_specs(specs, section, seen)
@@ -165,7 +165,7 @@ def parse_frozen_specs():
             try:
                 source = seen[frozenid]
             except KeyError:
-                source = FrozenSource.from_id(frozenid, pyfile)
+                source = FrozenSource.from_id(frozenid, suffix, pyfile)
                 seen[frozenid] = source
             else:
                 assert not pyfile or pyfile == source.pyfile, item
@@ -273,11 +273,11 @@ def _parse_spec(spec, knownids=None, section=None):
 class FrozenSource(namedtuple('FrozenSource', 'id pyfile frozenfile deepfreezefile')):
 
     @classmethod
-    def from_id(cls, frozenid, pyfile=None):
+    def from_id(cls, frozenid, suffix, pyfile=None):
         if not pyfile:
             pyfile = os.path.join(STDLIB_DIR, *frozenid.split('.')) + '.py'
             #assert os.path.exists(pyfile), (frozenid, pyfile)
-        frozenfile = resolve_frozen_file(frozenid, FROZEN_MODULES_DIR)
+        frozenfile = resolve_frozen_file(frozenid, FROZEN_MODULES_DIR, suffix)
         return cls(frozenid, pyfile, frozenfile, STDLIB_DIR)
 
     @classmethod
@@ -313,7 +313,7 @@ class FrozenSource(namedtuple('FrozenSource', 'id pyfile frozenfile deepfreezefi
         return self.id in BOOTSTRAP
 
 
-def resolve_frozen_file(frozenid, destdir):
+def resolve_frozen_file(frozenid, destdir, suffix):
     """Return the filename corresponding to the given frozen ID.
 
     For stdlib modules the ID will always be the full name
@@ -326,7 +326,7 @@ def resolve_frozen_file(frozenid, destdir):
             raise ValueError(f'unsupported frozenid {frozenid!r}')
     # We use a consistent naming convention for all frozen modules.
     frozen_symbol = FrozenSource.resolve_symbol(frozenid)
-    frozenfile = f"Frozen{frozen_symbol}.bin"
+    frozenfile = f"Frozen{frozen_symbol}.{suffix}"
 
     if not destdir:
         return frozenfile
@@ -636,11 +636,17 @@ def main():
     STDLIB_DIR = os.path.abspath(parsed_args.python_lib)
     FROZEN_MODULES_DIR = os.path.abspath(parsed_args.binary_dir)
 
-    # create module specs
-    modules = list(parse_frozen_specs())
+    if __graalpython__.is_bytecode_dsl_interpreter:
+        suffix = "bin_dsl"
+        assert os.path.isdir(parsed_args.binary_dir), "Frozen modules for the DSL should be built after the manual bytecode interpreter."
+    else:
+        suffix = "bin"
+        shutil.rmtree(parsed_args.binary_dir, ignore_errors=True)
+        os.makedirs(parsed_args.binary_dir)
 
-    shutil.rmtree(parsed_args.binary_dir, ignore_errors=True)
-    os.makedirs(parsed_args.binary_dir)
+    # create module specs
+    modules = list(parse_frozen_specs(suffix))
+
     # write frozen module binary files containing the byte code and class files
     # used for importing the binary files
     for src in _iter_sources(modules):
